requirements.txt
# Backend
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart  # For file uploads (face images)
supabase
python-dotenv
psycopg2-binary

# Client / Computer Vision
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
requests>=2.31.0  # For client-server communication
